pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pathly-adapters"
version = "1.0.0"
description = "Pathly adapter installer — stitches and deploys agent files to AI host tools."
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=8", "build", "twine"]

[project.scripts]
pathly-setup = "install_cli.__main__:main"
pathly-tokens = "pathly_telemetry.report:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = ["-p", "no:cacheprovider"]

[tool.setuptools.packages.find]
include = ["install_cli*", "pathly_telemetry*", "pathly_data*"]

[tool.setuptools.package-data]
"pathly_data" = [
    "core/**/*",
    "adapters/**/*",
    "adapters/claude/.claude-plugin/*",
    "adapters/codex/.codex-plugin/*",
    "schemas/*.json",
]
